requirements.txt
git+https://github.com/Rapptz/discord.py
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
discord.py==1.3.2
idna==2.9
isort==4.3.21
lavalink==3.0.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pylint==2.4.4
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
git+https://github.com/Rapptz/discord-ext-menus
git+https://github.com/Rapptz/discord.py@master#egg=discord.py[voice]
git+https://github.com/Devoxin/Lavalink.py@dev